@@ -7,7 +7,7 @@ package fspath
 import (
 	"io/fs"
 	"os"
-	"path/filepath"
+	"path"
 )
 
 // FS implements the fs interface and can also show a path where the fs is located.
@@ -26,7 +26,7 @@ type fsDir struct {
 
 // Path returns a path for the given names, based on the location of the file system.
 func (fs *fsDir) Path(names ...string) string {
-	return filepath.Join(append([]string{fs.path}, names...)...)
+	return path.Join(append([]string{fs.path}, names...)...)
 }
 
 // DirFS returns a file system for a directory, it keeps the path to implement the FS interface.
