@@ -17,7 +17,7 @@ import (
 	"github.com/pkg/errors"
 	"gopkg.in/yaml.v3"
 
-	"github.com/elastic/package-spec/code/go/internal/fspath"
+	"github.com/elastic/package-spec/v2/code/go/internal/fspath"
 )
 
 // File represents a file in the package.
