@@ -7,8 +7,8 @@ package semantic
 import (
 	"fmt"
 
-	"github.com/elastic/package-spec/code/go/internal/errors"
-	"github.com/elastic/package-spec/code/go/internal/fspath"
+	"github.com/elastic/package-spec/v2/code/go/internal/errors"
+	"github.com/elastic/package-spec/v2/code/go/internal/fspath"
 )
 
 // ValidateFieldGroups verifies if field groups don't have units and metric types defined.
