@@ -11,8 +11,8 @@ import (
 
 	"github.com/Masterminds/semver/v3"
 
-	ve "github.com/elastic/package-spec/code/go/internal/errors"
-	"github.com/elastic/package-spec/code/go/internal/fspath"
+	ve "github.com/elastic/package-spec/v2/code/go/internal/errors"
+	"github.com/elastic/package-spec/v2/code/go/internal/fspath"
 )
 
 var (
@@ -89,11 +89,9 @@ func validatePrereleaseTag(tag string) error {
 		}
 	}
 
-	return ve.ValidationErrors{
-		fmt.Errorf("prerelease tag (%s) should be one of [%s], or one of [%s] followed by numbers",
-			tag,
-			strings.Join(literalPrereleases, ", "),
-			strings.Join(numberedPrereleases, ", "),
-		),
-	}
+	return fmt.Errorf("prerelease tag (%s) should be one of [%s], or one of [%s] followed by numbers",
+		tag,
+		strings.Join(literalPrereleases, ", "),
+		strings.Join(numberedPrereleases, ", "),
+	)
 }
